+++ src/constants.ts
@@ -1,5 +1,8 @@
 export const MODULE_ID = "pneuma-payouts";
 
 export const PAYOUT_LEDGER_SETTING = "payoutLedger";
+export const DISCORD_LINKS_SETTING = "discordLinks";
+export const PAYOUT_JOURNAL_ID_SETTING = "payoutJournalId";
+export const PAYOUT_JOURNAL_DATA_SETTING = "payoutJournalData";
 
 export const PAYOUT_SCHEMA_VERSION = 1;