+++ src/main.ts
@@ -1,7 +1,12 @@
 import "./styles/pneuma-payouts.css";
 import { pneumaPayoutsApi } from "./api";
 import { MODULE_ID } from "./constants";
+import { registerDiscordLinks } from "./discord-summary";
 import { registerPayoutLedger } from "./payout-ledger";
+import {
+  ensurePayoutJournal,
+  registerPayoutJournalSettings,
+} from "./payout-journal";
 import { registerHumanityPromptHandler } from "./humanity-prompts";
 import { registerPayoutWindowControl } from "./window-controls";
 
@@ -14,6 +19,8 @@ Hooks.once("init", () => {
   console.info(`${MODULE_ID} | Initializing`);
 
   registerPayoutLedger();
+  registerDiscordLinks();
+  registerPayoutJournalSettings();
 
   const module = game.modules.get(MODULE_ID);
   if (module) module.api = pneumaPayoutsApi;
@@ -21,4 +28,5 @@ Hooks.once("init", () => {
 
 Hooks.once("ready", () => {
   console.info(`${MODULE_ID} | Ready`);
+  if (game.user?.isGM) void ensurePayoutJournal();
 });